requirements.txt
einops~=0.8.0
lsq_ellipse~=2.2.1
matplotlib~=3.6.0
numpy~=1.23.4
opencv_python~=4.6.0.66
scipy~=1.13.0
seaborn~=0.13.2
tqdm~=4.63.0
